update port name to use ip
CNM will end up renaming the interface provided to the
runtime, so instead of passing details via the interface,
provide a unique name for the port.  The container runtime can
then check the filesystem for this particular named file,
particularly for the appended IP against what is provided on the
interface.

Signed-off-by: Eric Ernst <eric.ernst@intel.com>